# docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.AddressBookForm:
    properties:
      alias:
        type: string
      force_always_relay:
        type: boolean
      hash:
        type: string
      hostname:
        type: string
      id:
        type: string
      login_name:
        type: string
      online:
        type: boolean
      password:
        type: string
      platform:
        type: string
      rdp_port:
        type: string
      rdp_username:
        type: string
      row_id:
        type: integer
      same_server:
        type: boolean
      tags:
        items:
          type: string
        type: array
      user_id:
        type: integer
      username:
        type: string
    required:
    - id
    type: object
...
  admin.GroupForm:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
  model.AddressBook:
    properties:
      alias:
        type: string
      created_at:
        type: string
      forceAlwaysRelay:
        type: boolean
      hash:
        type: string
      hostname:
        type: string
      id:
        type: string
      loginName:
        type: string
      online:
        type: boolean
      password:
        type: string
      platform:
        type: string
      rdpPort:
        type: string
      rdpUsername:
        type: string
      row_id:
        type: integer
      sameServer:
        type: boolean
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  model.AddressBookList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AddressBook'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.Group:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: integer
      updated_at:
        type: string
    type: object
  model.GroupList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Group'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
  /admin/address_book/create:
    post:
      consumes:
      - application/json
      description: 创建地址簿
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBook'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建地址簿
      tags:
      - 地址簿
  /admin/address_book/delete:
    post:
      consumes:
      - application/json
      description: 地址簿删除
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿删除
      tags:
      - 地址簿
  /admin/address_book/detail/{id}:
    get:
      consumes:
      - application/json
      description: 地址簿详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBook'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿详情
      tags:
      - 地址簿
  /admin/address_book/list:
    get:
      consumes:
      - application/json
      description: 地址簿列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 用户id
        in: query
        name: user_id
        type: integer
      - description: 是否是我的
        in: query
        name: is_my
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿列表
      tags:
      - 地址簿
  /admin/address_book/update:
    post:
      consumes:
      - application/json
      description: 地址簿编辑
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBook'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿编辑
      tags:
      - 地址簿
  /admin/file/oss_token:
    get:
      consumes:
      - application/json
      description: 获取ossToken
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 获取ossToken
      tags:
      - 文件
  /admin/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到本地
      parameters:
      - description: 上传文件示例
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 上传文件到本地
      tags:
      - 文件
  /admin/group/create:
    post:
      consumes:
      - application/json
      description: 创建群组
      parameters:
      - description: 群组信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.GroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Group'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建群组
      tags:
      - 群组
  /admin/group/delete:
    post:
      consumes:
      - application/json
      description: 群组删除
      parameters:
      - description: 群组信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.GroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 群组删除
      tags:
      - 群组
  /admin/group/detail/{id}:
    get:
      consumes:
      - application/json
      description: 群组详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Group'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 群组详情
      tags:
      - 群组
  /admin/group/list:
    get:
      consumes:
      - application/json
      description: 群组列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.GroupList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 群组列表
      tags:
      - 群组
  /admin/group/update:
    post:
      consumes:
      - application/json
      description: 群组编辑
      parameters:
      - description: 群组信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.GroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Group'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 群组编辑
      tags:
      - 群组
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"